Platformio ini restructure and examples folder (#29)
* Restructure platformio ini to have a examples folder

* Solve merging issues
