+++ platformio.ini
@@ -25,6 +25,9 @@ default_envs =
   lolin32
 ;  esp32doit-devkit-v1
 
+# Modify this line to build and upload examples from the examples folder e.g. examples/dmn/
+src_dir = src
+
 [common]
 build_flags_ext =
 
@@ -51,59 +54,49 @@ lib_deps_32 =
   HTTPClient
   AsyncTCP
 
-[env:d1_mini]
-platform = espressif8266
+[env]
 framework = arduino
-board = d1_mini
-board_build.filesystem = littlefs
 build_flags =
   ${common.build_flags_ext}
 lib_deps =
   ${common.lib_deps_builtin}
   ${common.lib_deps_external}
+
+[env:d1_mini]
+platform = espressif8266
+board = d1_mini
+board_build.filesystem = littlefs
+lib_deps =
+  ${env.lib_deps}
   ${common.lib_deps_8266}
 
 [env:huzzah]
 platform = espressif8266
-framework = arduino
 board = huzzah
-build_flags =
-  ${common.build_flags_ext}
 lib_deps =
-  ${common.lib_deps_builtin}
-  ${common.lib_deps_external}
+  ${env.lib_deps}
   ${common.lib_deps_8266}
 
 [env:nodemcu]
 platform = espressif8266
-framework = arduino
 board = nodemcuv2
-build_flags =
-  ${common.build_flags_ext}
 lib_deps =
-  ${common.lib_deps_builtin}
-  ${common.lib_deps_external}
+  ${env.lib_deps}
   ${common.lib_deps_8266}
 
 [env:lolin32]
 platform = espressif32
-framework = arduino
 board = lolin32
 build_flags =
-  ${common.build_flags_ext}
+  ${env.build_flags}
   -DBOARD_LOLIN_D32
 lib_deps =
-  ${common.lib_deps_builtin}
-  ${common.lib_deps_external}
+  ${env.lib_deps}
   ${common.lib_deps_32}
 
 [env:esp32doit-devkit-v1]
 platform = espressif32
-framework = arduino
 board = esp32doit-devkit-v1
-build_flags =
-  ${common.build_flags_ext}
 lib_deps =
-  ${common.lib_deps_builtin}
-  ${common.lib_deps_external}
+  ${env.lib_deps}
   ${common.lib_deps_32}
